Divide vouchers page hooks
