+++ idea/frontend/src/pages/vouchers/hooks/index.ts
@@ -1,0 +1,4 @@
+import { useVoucherFilters } from './use-voucher-filters';
+import { useSearchQuery } from './use-search-query';
+
+export { useVoucherFilters, useSearchQuery };
